Semistandardized README
This is an interesting project, because it isnt just a code package. As such, Standard README may not be the best for it. Still, I collapsed some sections, made it a bit easier to read. I think directing people to the readme.io docs makes a lot of sense - what you have here is basically extra information. Some of it is important to state twice, some of it less so.

Who is the code licensed under? That is not clear.

I hope this helps!